refactor(orchestrator): align sampler module name
