+++ src/prime_rl/orchestrator/curriculum/samplers/__init__.py
@@ -2,6 +2,6 @@
 
 from prime_rl.orchestrator.curriculum.samplers.base import TaskSampler
 from prime_rl.orchestrator.curriculum.samplers.pool import DifficultyPoolSampler
-from prime_rl.orchestrator.curriculum.samplers.simple import StandardSampler
+from prime_rl.orchestrator.curriculum.samplers.standard import StandardSampler
 
 __all__ = ["DifficultyPoolSampler", "StandardSampler", "TaskSampler"]
